API設計書に基づくテスト修正(もくもく会一覧部分)
