+++ backend/routes/api.php
@@ -52,9 +52,4 @@ Route::namespace('App\Http\Controllers')->group(function () {
     Route::prefix('search')->name('search')->group(function () {
         Route::get('', 'SearchController@index');
     });
-    Route::namespace('Master')->group(function () {
-        Route::get('/get_parties', 'PartyController@index')->name('parties');
-        Route::get('/get_created_parties', 'PartyController@indexCreated')->name('created_parties');
-        Route::get('/get_participated_parties', 'PartyController@indexParticipated')->name('participated_parties');
-    });
 });
